dom/xbl/test/test_fieldScopeChain.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1095660
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug  **/
  SimpleTest.waitForExplicitFinish();
  window.baz = 1;
  document.baz = 2;
  addLoadEvent(function() {
    is(document.querySelector("pre").bar, 2,
       "Should have document on field scope chain");
    SimpleTest.finish();
  });
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1095660">Mozilla Bug </a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test" style="-moz-binding: url(file_fieldScopeChain.xml#foo)">
</pre>
</body>
</html>
